Add core galvo device property. Progress on Rapp. Simplify some core code.
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@8555 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ MMCore/CircularBuffer.cpp
@@ -253,7 +253,7 @@ const ImgBuffer* CircularBuffer::GetNthFromTopImageBuffer(unsigned long n) const
    if (n >= frameArray_.size() )
       return 0;
 
-   if (insertIndex_ <= n)
+   if ((unsigned long) insertIndex_ <= n)
    {
       return frameArray_[(frameArray_.size() - n + insertIndex_ - 1) % frameArray_.size()].FindImage(0, 0);
    }
